client/requirements.txt
ttkbootstrap
requests
Pillow
PyJWT